backend tests fixed

+++ ngraph/test/backend/adaptive_max_pool.in.cpp
@@ -16,7 +16,7 @@ using TestEngine = test::ENGINE_CLASS_NAME(${BACKEND_NAME});
 NGRAPH_TEST(${BACKEND_NAME}, adaptive_max_pool_1d)
 {
     auto data = make_shared<op::Parameter>(element::f32, Shape{2, 3, 7});
-    auto output_shape = op::Constant::create<int64_t>(element::i64, Shape{1}, {3});
+    auto output_shape = op::Constant::create<int32_t>(element::i32, Shape{1}, {3});
     auto adaptive_pool = make_shared<op::v8::AdaptiveMaxPool>(data, output_shape, element::i32);
     auto fun = make_shared<Function>(adaptive_pool->outputs(), ParameterVector{data});
 
@@ -34,7 +34,7 @@ NGRAPH_TEST(${BACKEND_NAME}, adaptive_max_pool_1d)
                                         3,  4,  1,
                                         2, -2, -2,
                                         3,  2,  2};
-    std::vector<int64_t> expected_indices{1, 3, 4,
+    std::vector<int32_t> expected_indices{1, 3, 4,
                                           1, 3, 6,
                                           1, 4, 4,
 
@@ -44,14 +44,14 @@ NGRAPH_TEST(${BACKEND_NAME}, adaptive_max_pool_1d)
     auto test_case = test::TestCase<TestEngine>(fun);
     test_case.add_input<float>(Shape{2, 3, 7}, inputs);
     test_case.add_expected_output<float>(Shape{2, 3, 3}, expected_result);
-    test_case.add_expected_output<int64_t>(Shape{2, 3, 3}, expected_indices);
+    test_case.add_expected_output<int32_t>(Shape{2, 3, 3}, expected_indices);
     test_case.run();
 }
 
 NGRAPH_TEST(${BACKEND_NAME}, adaptive_max_pool_2d)
 {
     auto data = make_shared<op::Parameter>(element::f32, Shape{1, 3, 7, 10});
-    auto output_shape = op::Constant::create<int64_t>(element::i64, Shape{2}, {3, 3});
+    auto output_shape = op::Constant::create<int32_t>(element::i32, Shape{2}, {3, 3});
     auto adaptive_pool = make_shared<op::v8::AdaptiveMaxPool>(data, output_shape, element::i32);
     auto fun = make_shared<Function>(adaptive_pool->outputs(), ParameterVector{data});
 
@@ -89,7 +89,7 @@ NGRAPH_TEST(${BACKEND_NAME}, adaptive_max_pool_2d)
                                        4, 3, 4,
                                        4, 3, 3,
                                        4, 4, 4};
-    std::vector<int64_t> expected_indices{22, 5 , 16,
+    std::vector<int32_t> expected_indices{22, 5 , 16,
                                           22, 43, 48,
                                           43, 43, 48,
 
@@ -104,14 +104,14 @@ NGRAPH_TEST(${BACKEND_NAME}, adaptive_max_pool_2d)
     auto test_case = test::TestCase<TestEngine>(fun);
     test_case.add_input<float>(Shape{1, 3, 7, 10}, inputs);
     test_case.add_expected_output<float>(Shape{1, 3, 3, 3}, expected_result);
-    test_case.add_expected_output<int64_t>(Shape{1, 3, 3, 3}, expected_indices);
+    test_case.add_expected_output<int32_t>(Shape{1, 3, 3, 3}, expected_indices);
     test_case.run();
 }
 
 NGRAPH_TEST(${BACKEND_NAME}, adaptive_max_pool_3d)
 {
     auto data = make_shared<op::Parameter>(element::f32, Shape{2, 2, 3, 3, 3});
-    auto output_shape = op::Constant::create<int64_t>(element::i64, Shape{3}, {2, 2, 2});
+    auto output_shape = op::Constant::create<int32_t>(element::i32, Shape{3}, {2, 2, 2});
     auto adaptive_pool = make_shared<op::v8::AdaptiveMaxPool>(data, output_shape, element::i32);
     auto fun = make_shared<Function>(adaptive_pool->outputs(), ParameterVector{data});
 
@@ -141,7 +141,7 @@ NGRAPH_TEST(${BACKEND_NAME}, adaptive_max_pool_3d)
 
                                        3, 1, 4, 4,
                                        3, 3, 4, 3};
-    std::vector<int64_t> expected_indices{4 , 4 , 4 , 4 ,
+    std::vector<int32_t> expected_indices{4 , 4 , 4 , 4 ,
                                           22, 22, 24, 22,
 
                                           3 , 14, 3 , 8 ,
@@ -156,6 +156,6 @@ NGRAPH_TEST(${BACKEND_NAME}, adaptive_max_pool_3d)
     auto test_case = test::TestCase<TestEngine>(fun);
     test_case.add_input<float>(Shape{2, 2, 3, 3, 3}, inputs);
     test_case.add_expected_output<float>(Shape{2, 2, 2, 2, 2}, expected_result);
-    test_case.add_expected_output<int64_t>(Shape{2, 2, 2, 2, 2}, expected_indices);
+    test_case.add_expected_output<int32_t>(Shape{2, 2, 2, 2, 2}, expected_indices);
     test_case.run();
 }
